backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
cryptography==41.0.7
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-generativeai==0.8.3
pypdf==4.2.0
